deleted not needed files and added folder structure for iteration_1. added instructions on the readme file
