+++ src/Iteration_1/src/TrainConnection.java
@@ -1,3 +1,4 @@
+package Iteration_1.src;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
