--- src/Train_Search.java
@@ -1,4 +1,0 @@
-public class Train_Search {
-
-    
-}
